D2 leg 1: a cell reads a device register through a granted BAR window
A cell could not reach a device's registers at all: `SYS_GRANT` refuses
`MemKind::DeviceBar`, and nothing else mapped one. That refusal is right and
stays - a cell must not be able to give itself a device - but it left the
*mapping* mechanism missing too, which is the first of the three legs
DRIVERS.md D2 needs before a driver can live in a cell.

`load::map_device_bar` maps a whole-page span of an enumerated, assigned,
memory-space BAR into a cell's address space as the new `MapPerm::UserDevice`:
uncached device memory per ISA - x86-64 `PCD|PWT` (PAT entry 3 = UC), ARM64
MAIR attr 0 (Device-nGnRnE, unshared), and on RISC-V a plain mapping, because
base Sv39 has no cacheability field at all and claiming an attribute the page
table format cannot express would be a lie. It refuses I/O-space BARs,
unassigned BARs and anything over `USER_BAR_MAX` (4 MiB), and places the window
in its own fixed region (`USER_BAR_VA`, 28 GiB) below every ISA's user floor.

The authority is launcher-minted, like the W^X exception, the cell-spawn
capability and the queue pair: the cell asks for nothing and can ask for
nothing.

Proven by `nvmefs` on all three ISAs against a value the cell cannot
fabricate - the NVMe controller's VS register at BAR0+0x08, read at the
unprivileged level through the granted window and equal to the kernel's own
MMIO read of the same register (0x10400 under this QEMU: a live reading, not a
constant). Two controls, both observed failing: aiming the cell 4 bytes off
breaks the comparison, and removing the grant entirely faults - so the read is
a real access through the page tables rather than a value the cell could have
produced without a mapping. The window's bound is asserted permanently in the
same phase, because a mapping that overran its BAR would hand the cell the
next device's registers: the same cell aimed one page past the mapping faults.

Honest: cacheability is unobservable here (TCG models no cache), so the
attributes are asserted by construction in the per-ISA paging code and not by
measurement. And this is leg 1 of 3 - a config-space capability, a
cell-lifetime IOMMU domain, a cell-reachable DMA-map verb and the per-device
IRQ wait are not built, so no driver has been re-homed into a cell. D2 stays
open; docs/DRIVERS.md 4.1 and 8 say so.

Also fixes two lint regressions from the previous slice (an unnecessary cast
in the Linux ENOENT logger, an argument-count warning on the runtime-disk
fixture builder).

66/66 test kernels on all three ISAs.

+++ kernel/src/linux/mod.rs
@@ -449,7 +449,7 @@ fn trace_record(nr_val: u64, ctl: &Ctl) {
     // dozen, so an unbounded log would bury the signal it exists to surface. First
     // [`ENOENT_LOG_MAX`] only, `open` excluded because it already reports itself with
     // more information than a number.
-    if ret == -(errno::ENOENT as i64) && !is_open_nr(nr_val) {
+    if ret == -errno::ENOENT && !is_open_nr(nr_val) {
         let n = unsafe { *addr_of!(ENOENT_LOGGED) };
         if n < ENOENT_LOG_MAX {
             unsafe { *addr_of_mut!(ENOENT_LOGGED) = n + 1 };
